requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
lxml==5.3.0
PyPDF2==3.0.1
python-docx==1.1.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
